feat: extract common json, split edit and manage edit tests
